requirements.txt
# Dependencias Nucleares Symbiomemesis
pandas
numpy
pydantic>=2.0.0
tapeagents
python-dotenv
litellm